fix: Fixed permissions with admins and editing other admin roles

+++ src/actions/user/roles/deleteUserRole.ts
@@ -1,7 +1,8 @@
 'use server'
 
 import { redirectIfInvalidSession } from "@/auth/auth"
-import { RoleTypes } from "@/auth/roles/Roles"
+import getOrgRole from "@/auth/roles/getOrgRole"
+import { getRoleFromID, RoleTypes } from "@/auth/roles/Roles"
 import { throwIfInsufficientPerms } from "@/auth/roles/throwIfInsufficientPerms"
 import { prisma } from "@/database/prisma"
 import { revalidatePath } from "next/cache"
@@ -16,7 +17,13 @@ export default async function deleteUserRole(userUUID: string, orgUUID: string) 
     await redirectIfInvalidSession()
     await throwIfInsufficientPerms(RoleTypes.Admin)
 
-    await prisma.role.deleteMany({ where: { userId: userUUID, organizationId: orgUUID }})
+    const currentRole = await prisma.role.findFirst({ where: { userId: userUUID, organizationId: orgUUID } })
+    const userRole = await getOrgRole()
+    if (userRole.level <= getRoleFromID(currentRole?.role ?? "").level) { // You cannot edit the roles of other admins 
+        throw new Error("Insufficient Permissions")
+    }
+
+    await prisma.role.deleteMany({ where: { userId: userUUID, organizationId: orgUUID } })
 
     revalidatePath("/user/users/" + userUUID)
 }
